implement omniauth logic
