+++ .byebug_history
@@ -1,3 +1,9 @@
+exit
+current_user
+env["omniauth.auth"]
+request.env["omniauth.auth"]
+request.env
+env["omniauth.auth"]
 c
 user_facebook_omniauth_authorize_path
 user_omniauth_authorize_path(:facebook)
