+++ app/controllers/application_controller.rb
@@ -3,6 +3,7 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
+  # before_action :ensure_signup_complete, only: [:new, :create, :update, :destroy]
   before_action :configure_devise_permitted_parameters, if: :devise_controller?
 
   protected
@@ -20,4 +21,15 @@ class ApplicationController < ActionController::Base
       }
     end
   end
+
+  # def ensure_signup_complete
+  #   # Ensure we don't go into an infinite loop
+  #   return if action_name == 'finish_signup'
+
+  #   # Redirect to the 'finish_signup' page if the user
+  #   # email hasn't been verified yet
+  #   if current_user && !current_user.email_verified?
+  #     redirect_to finish_signup_path(current_user)
+  #   end
+  # end
 end
