+++ app/views/users/finish_signup.html.haml
@@ -1,0 +1,14 @@
+#add-email{:class => "container"}
+  %h1 Add Email
+  = form_for(current_user, :as => 'user', :url => finish_signup_path(current_user), :html => { role: 'form'}) do |f|
+    - if @show_errors && current_user.errors.any?
+      #error_explanation
+        - current_user.errors.full_messages.each do |msg|
+          = msg
+    .form-group
+      = f.label :email
+      .controls
+        = f.text_field :email, :autofocus => true, :value => '', class: 'form-control input-lg', placeholder: 'Example: email@me.com'
+        %p{:class => "help-block"}Please confirm your email address. No spam.
+    .actions
+      = f.submit 'Continue', :class => 'btn btn-primary'
