+++ config/application.rb
@@ -6,7 +6,7 @@ require 'rails/all'
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
-module TestSimple
+module JobHero
   class Application < Rails::Application
 
     config.generators do |g|
